+++ ebird/tests/integration/test_product.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+
+from unittest import TestCase
+
+from ebird import region_notable, hotspot_summary
+
+
+class ProductIntegrationTests(TestCase):
+    """Tests for the product-related end-points in the eBird API."""
+
+    @classmethod
+    def setUpClass(cls):
+        records = region_notable('US-NY', hotspot=True)
+        cls.location = records[0]['locID']
+
+    def test_hotspot_summary(self):
+        hotspot_summary(self.location)