+++ ebird/tests/unit/test_find_regions.py
@@ -7,16 +7,16 @@ try:
 except ImportError:
     import unittest.mock as mock
 
-from ebird.core import find_locations, FIND_LOCATIONS_URL
+from ebird.reference import FIND_REGIONS_URL, find_regions
 
 
 # noinspection PyUnusedLocal
-def get_locations(url, params):
-    pass
+def get_content(url, params):
+    return '[]'
 
 
-class FindLocationsTests(TestCase):
-    """Tests for the find_locations() API call."""
+class FindRegionsTests(TestCase):
+    """Tests for the find_regions() API call."""
 
     parameters = [
         (('bcr',  'Coastal'), {'rtype': 'bcr', 'match': 'Coastal'}),
@@ -30,25 +30,25 @@ class FindLocationsTests(TestCase):
         ('other', 'west'),
     ]
 
-    @mock.patch('ebird.core.get_locations', side_effect=get_locations)
+    @mock.patch('ebird.reference.get_content', side_effect=get_content)
     def test_url(self, mocked_function):
         """Verify the correct URL is used to fetch the records."""
-        find_locations('country', 'North')
+        find_regions('country', 'North')
         actual = mocked_function.call_args[0][0]
-        self.assertEqual(FIND_LOCATIONS_URL, actual)
+        self.assertEqual(FIND_REGIONS_URL, actual)
 
-    @mock.patch('ebird.core.get_locations', side_effect=get_locations)
+    @mock.patch('ebird.reference.get_content', side_effect=get_content)
     def test_parameters(self, mocked_function):
         """Verify only non-default values for query string parameters are sent."""
         for idx, (args, expected) in enumerate(self.parameters):
-            find_locations(*args)
+            find_regions(*args)
             actual = mocked_function.call_args[0][1]
             self.assertDictEqual(expected, actual, msg="Pattern %d failed" % idx)
 
     # noinspection PyUnusedLocal
-    @mock.patch('ebird.core.get_locations', side_effect=get_locations)
+    @mock.patch('ebird.reference.get_content', side_effect=get_content)
     def test_validation(self, mocked_function):
         """Verify the function arguments are validated."""
         for idx, args in enumerate(self.validation):
             with self.assertRaises(ValueError, msg='Pattern %d failed' % idx):
-                find_locations(*args)
+                find_regions(*args)