+++ ebird/tests/unit/test_geo_observations.py
@@ -7,12 +7,12 @@ try:
 except ImportError:
     import unittest.mock as mock
 
-from ebird.core import geo_observations, GEO_OBSERVATIONS_URL
+from ebird.data import GEO_OBSERVATIONS_URL, geo_observations
 
 
 # noinspection PyUnusedLocal
-def get_observations(url, params):
-    pass
+def get_content(url, params):
+    return '[]'
 
 
 class GeoObservationsTests(TestCase):
@@ -80,14 +80,14 @@ class GeoObservationsTests(TestCase):
         # TODO Currently do not have validation tests for hotspot
     ]
 
-    @mock.patch('ebird.core.get_observations', side_effect=get_observations)
+    @mock.patch('ebird.data.get_content', side_effect=get_content)
     def test_url(self, mocked_function):
         """Verify the correct URL is used to fetch the records."""
         geo_observations(0.0, 0.0)
         actual = mocked_function.call_args[0][0]
         self.assertEqual(GEO_OBSERVATIONS_URL, actual)
 
-    @mock.patch('ebird.core.get_observations', side_effect=get_observations)
+    @mock.patch('ebird.data.get_content', side_effect=get_content)
     def test_parameters(self, mocked_function):
         """Verify only non-default values for query string parameters are sent."""
         for idx, (pattern, expected) in enumerate(self.parameters):
@@ -96,7 +96,7 @@ class GeoObservationsTests(TestCase):
             self.assertDictEqual(expected, actual, msg="Pattern %d failed" % idx)
 
     # noinspection PyUnusedLocal
-    @mock.patch('ebird.core.get_observations', side_effect=get_observations)
+    @mock.patch('ebird.data.get_content', side_effect=get_content)
     def test_validation(self, mocked_function):
         """Verify the function arguments are validated."""
         for idx, pattern in enumerate(self.validation):