+++ ebird/tests/unit/test_list_regions.py
@@ -7,16 +7,16 @@ try:
 except ImportError:
     import unittest.mock as mock
 
-from ebird.core import list_locations, LIST_LOCATIONS_URL
+from ebird.reference import LIST_REGIONS_URL, list_regions
 
 
 # noinspection PyUnusedLocal
-def get_locations(url, params):
-    pass
+def get_content(url, params):
+    return '[]'
 
 
-class ListLocationsTests(TestCase):
-    """Tests for the list_locations() API call.
+class ListRegionsTests(TestCase):
+    """Tests for the list_regions() API call.
 
     Since the function only has 2 arguments (one positional, one
     keyword) the data patterns can be simplified if we treat the
@@ -45,25 +45,25 @@ class ListLocationsTests(TestCase):
         ('other', 'US'),
     ]
 
-    @mock.patch('ebird.core.get_locations', side_effect=get_locations)
+    @mock.patch('ebird.reference.get_content', side_effect=get_content)
     def test_url(self, mocked_function):
         """Verify the correct URL is used to fetch the records."""
-        list_locations('country')
+        list_regions('country')
         actual = mocked_function.call_args[0][0]
-        self.assertEqual(LIST_LOCATIONS_URL, actual)
+        self.assertEqual(LIST_REGIONS_URL, actual)
 
-    @mock.patch('ebird.core.get_locations', side_effect=get_locations)
+    @mock.patch('ebird.reference.get_content', side_effect=get_content)
     def test_parameters(self, mocked_function):
         """Verify only non-default values for query string parameters are sent."""
         for idx, (args, expected) in enumerate(self.parameters):
-            list_locations(*args)
+            list_regions(*args)
             actual = mocked_function.call_args[0][1]
             self.assertDictEqual(expected, actual, msg="Pattern %d failed" % idx)
 
     # noinspection PyUnusedLocal
-    @mock.patch('ebird.core.get_locations', side_effect=get_locations)
+    @mock.patch('ebird.reference.get_content', side_effect=get_content)
     def test_validation(self, mocked_function):
         """Verify the function arguments are validated."""
         for idx, args in enumerate(self.validation):
             with self.assertRaises(ValueError, msg='Pattern %d failed' % idx):
-                list_locations(*args)
+                list_regions(*args)