+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added function for fetching lists of species.
 - Added function for fetching nearby hotspots.
 - Added function for fetching lists of hotspots for a region.
+### Changed
+- Renamed list_locations() to list_regions() and find_locations() to 
+find_regions() as "locations" refer to specific sites in the functions 
+that fetch observations.
 
 ## [0.2.0] - 2017-02-18
 ### Added